pass information on Jenkins user to container launch script
Some container launch scripts may need to know the username and user/group IDs
of the Jenkins user which the actual build steps will run as.  In particular
the container launch script may want to provide a background service, such as
a graphical environment, for use during the build.

This change passes in this information as part of the container's process
environment.