remove unnecessary spam from the docker exec command line
Variable definitions passed into docker run via --env are also included
in the environment of commands launched via docker exec.  So for every
variable that matches the origin docker run environment, we can exclude
that value from the command line and prevent unnecessary value spam in
the build log as a bonus.