Adding ability to disable workspace mapping for running inside the container
This is useful when you have you build your own container and want to execute
a process that produces the output and you don't necessarily want to
operate on files from Jenkins' workspace. Our usecase is that we pre-build
our application tests as a container and want to execute mvn inside of it.
We don't want to have to specify pom.xml path in the existing mvn command.