JENKINS-66171: Get built docker image ID using --iidfile
Previously we were capturing `docker build` output, which led to some problems:
1. Other programs could print similar strings, which would match but did not
contain the image ID. This was fixed by taking the last capture as image ID.
2. `docker build` output has changed in version 23, so now we'd have to match
the output twice. First pass for docker < 23, and another for >= 23

Docker has had the `--iidfile` option for some time now, so switch to using it.
It is a much less fragile way to capture the image ID.

Tested with Docker 20.10.23 and 23.0.1